extract v2_client.py module
